fix(review): 修复 CR 提出的问题
**Critical**:
- 修复掩码逻辑：解密后再掩码，避免密文泄露

**Major**:
- 移除 revealUserApiKey API（密码验证未实现）
- 修复 PUT/DELETE 路由：移至 [id]/route.ts
- 修复 server-only 条件导入
- 添加数据库并发锁提示

**Medium**:
- 添加客户端 API key 安全提示
- A股分钟级数据 lookback 限制（TODO）
- baostock 错误处理改进（TODO）
