+++ apps/dashboard/src/app/api/user/settings/route.ts
@@ -76,51 +76,3 @@ export async function POST(request: NextRequest) {
     );
   }
 }
-
-/**
- * PUT /api/user/settings/[id]
- * 更新指定配置。
- *
- * Body: { provider?, model?, api_key?, custom_base_url?, custom_provider_name?, label? }
- */
-export async function PUT(
-  request: NextRequest,
-  { params }: { params: { id: string } },
-) {
-  try {
-    const subject = await getSessionSubject();
-    const body = await request.json();
-
-    await updateUserLLMConfig(subject, params.id, body);
-
-    return NextResponse.json({ success: true });
-  } catch (err) {
-    console.error("[PUT /api/user/settings]", err);
-    return NextResponse.json(
-      { error: err instanceof Error ? err.message : "Unknown error" },
-      { status: 500 },
-    );
-  }
-}
-
-/**
- * DELETE /api/user/settings/[id]
- * 删除指定配置。
- */
-export async function DELETE(
-  request: NextRequest,
-  { params }: { params: { id: string } },
-) {
-  try {
-    const subject = await getSessionSubject();
-    await deleteUserLLMConfig(subject, params.id);
-
-    return NextResponse.json({ success: true });
-  } catch (err) {
-    console.error("[DELETE /api/user/settings]", err);
-    return NextResponse.json(
-      { error: err instanceof Error ? err.message : "Unknown error" },
-      { status: 500 },
-    );
-  }
-}
